add templates, EB config file, and more
-options.config
this is for AWS EB. it sets environment variables for all the instances that will be created.

-application.py
created some new routes for 404 & 505 errors.

the geckodriver (without exe) is for linux, which is what AWS uses.

-templates
just html stuff mostly.

-testing.py
the if opsys =='nt' statement there is to switch the version of the web driver, depending on if we're working in dev or production. in dev, I'm on windows. in prod it's on AWS, which runs linux. two different binaries needed.

the rest is self explanatory