add second hour, fix div display issues
adding support to display the upcoming hour's booth schedule. I think this will be more useful than only showing the current hour. you'll know who's supposed to be in there now and who's supposed to be in there shortly. that's the main update for this commit. also:

proxy.conf
nginx was throwing a strange error about writing an upstream to a file, and apparently it was because it was running out of memory, or something. this should help if that occurs again.

-delete.py
was solely was tested and not needed

-load images (TL & NPL logos) locally instead of loading over internet

-finally got the divs to center themselves correctly and be responsive depending on screen size!

testing.py
need to rename this file!! anyways, this is where most of the work went. again, it's mostly about bringing in the VIC data for the next/upcoming hour so we can show that on the page.