fix variable scope, html/JS changes
-application.py
the current_hour variable was outside the function for the homepage route. this meant it was declared once and just sat there, which led to unexpected results. now it gets declared every time the route is called.

-misc
misc HTML/JS changes to some pages. adding a JS function to spit out the date at the top of the page. this div is going outside the jinga block, so it will display on every page (including on days we're closed.