new landing page, route for EB health, readme updates, more
-application.py
our new home route (/) now loads a landing/loading page which tells the user "just a moment...". this is because selenium can take 8+ seconds to do its thing, so a visual cue to the user that something is indeed happening in the background is useful.

also adding a route (/health) specifically/solely for AWS EB health checks. this is explained further in the readme file.

-readme
the previous method of circumventing the AWS EB health check turned out...not to work. this is a better method, at least for now.

also some other various cleanup in this file.

-static/misc.js
moving the main JS under a single, separate file in the static folder and linking to it from HTML.

-css
also moving all the generic CSS to a separate file. the dot elastic stuff is all for the new landing page.

-home.html
moving CSS & JS to separate files.

previously I had <h2> tags inside of <h1> tags, which is incorrect syntax. fixed here.

-land.html
new landing page, as explained above. PLEASE NOTE: for some reason if you run the refresh function as plain old JS, the page fails to load all the assets that appear after the script. so strange. I played around with it for a while, but finally settled on using the 'setTimeout' function, which delays implementing the script by x seconds. it's a hack, but it'll have to work for now.

-testing.py
I need to change the name of this file.

fixed "next hour" for 3pm and onward hours. if the real time is 3pm or later, the "next hour" block just shows "Rest of Day". previous formatting was confusing.