proper refreshing!!, real time clock
-misc.js
adding a real time clock to the top of the page.

-homt.html
adding a script at the bottom of page (need to move to separate file) for refreshing the page. This is a more proper way to do this, but still needs work. every 60.5 seconds it checks if we are at the top of the hour. if so, it refreshes the page. if not, if waits another 60.5 seconds and checks again. it's a little hacky to me, but works for now. reminder to self to move this script up into the main content block instead of using its own custom blocks like we're currently doing it.

plus changes for the real time clock.