fix version issue
see comment on line 106 of testing.py

basically, webdriver_manager always installs the latest version of the driver. however, recently the newest driver only supported the very newest version of the browser itself. BUT that version of the browser was in BETA mode. I don't understand how this happened, but just in case it is probably safe, for now, to list specifically which version of the driver we want to use. thankfully webdriver_manager has an easy-to-use attribute for this.