another fix for driver version
for some reason, use webdriver manager to install the latest driver is working just fine on our AWS linux server, but not on our windows dev machine (see previous comment). so this is a hacky way to deal with it, for now.

again, please see previous commit for reference and context.