increase driver cache range, other misc
-testing.py
webdriver manager defaults to checking for and downloading a new driver every 1 day. this is too often for our needs. thankfully they built in a changeable attribute for this. we're increasing it to 30 days. that should be plenty long enough.

-land.html
this is the landing page for the home route (/). slightly increasing the time (from .3 to .5 seconds) before the JS redirects the page. previously, the page would start redirecting sometimes before the images had loaded. really just need a better solution for this in general. this feels hacky.

-misc
adding alt text for TL & NPL logos.