fix initial JS redirect, fix booth 1 return case
-land.html

previously, the landing page was loading a JS redirect function after .5 seconds, so that the HTML, CSS, and images could all load to the page before it began redirecting, because once it starts redirecting, it stopped loading images. this looked goofy if all the images weren't yet loaded. now, we're using the built-in 'window.onload' functionality to wait until everything has loaded.

also, moving this function out of the special Jinga block and just into the above block. no need for a special block here. anything inside the main block will override the template.

-testing.py

previously, if we were running a check - only on the 2nd hour - on whether selenium was finding the hour and booth name on the page. if it did not find it, then we were assuming this meant the booth was closed. but we were not running this check for the 1st hour. this lead to problems, for example, on holidays, when there is neither. thus, flask was returning an error. now, if someone visits the site on a holiday, or if for another reason it's not there, it will show CLOSED for the 1st hour as well.

we're doing this check with the 'in locals()' functionality. basically, if a variable has been declared, it resides in 'locals' variable temporarily. the way the selenium script is currently written, if it cannot find that booth and that hour on the page, the variable is not being declared.

this needs to be improved! this feels hackish and not a good permanent solution.