fix ChromeDriver version for AWS linux
this continues to be a problem. first it was a problem on Windows, now it's a problem on AWS's version of Linux.

we're using Yum on AWS to download the Chrome browser. But apparently it's not installing a very recent copy of the browser. Meanwhile, Webdriver_Manager is downloading the very latest driver, which apparently doesn't support the newest browser. so we're declaring a specific version of the driver to download here.

however, this is not really a permanent solution and eventually Yum will start downloading a more recent version and we'll need to do this all over again.

GAH!