fix chrome driver
OK, so every time we deploy, there may or may not be a new version of Chromium released. if there IS a new version, then we need to specify the new driver to match as well. if we do not specify which driver to use, ChromeDriverManager will pick one, but it seems it alawys picked the wrong one. this is frustrating.

ideally we would be able to specify the version of the browser AND the version of the driver, so they match up.