fix styling for iPad
currently this site is being displayed on a 2nd generation iPad, whose screen is W: 768, H: 1024. The logos are too far down the page so that, when we're scrolled to the top of the page, only half of the logo is visible, which looks...bad. here is a CSS fix for that. this moves the logos up a bit ONLY for screens whose width is exactly 768 pixels.