refresh page twice per hour
previously, the page would refresh itself once per hour, at the top of the hour (during the first minute of the hour).

TL Staff requested we refresh more often, in case the VIC schedule changes at some point during the hour. so, now, in addition to the top of the hour, we're also refreshing at the bottom of the hour (when date().getMinutes() is 30).