fix for hiding banner message div
we're currently using a super old, Gen2 iPad. apparently it is using such an old version of Safari, that it doesn't support the element.remove() method, which is what we were using to hid the "banner" message div when there is in fact no banner message. so, for now, instead of writing an if statement to figure out how old the browser version is (since the iPad is temporary, we don't want to spend that much time on it), we're just going to use this other way of hiding the div.