new way of displaying banner
previously, we were trying to use JS to remove the banner div, if there was no message. however, our old iPad had issues with this method. long story short, now we're using jinja logic to decide whether to render the div on the page at all.

if the contents of the message.txt file is empty, we are declaring the value false. in jinja we're doing the check - if false, don't render the div at all, if anything but false, it will render.

now we don't have to worry about various JS methods not being supported across devices!