CSS for BrightSign
increasing the font slightly for easier legibility from a distance.

adding the second media query here. this is because, if only using the device width, any other screen with that width will use these values, and it turns out a whole bunch of normal screens have this resolution. on the BrightSign, we have touch/mouse/hover actions disabled, so with the combination of resolution and hover, this should target very few other devices.