update gitignore
since we're now using AWS CodePipeline for CD, we just need to set environment variables in ELB and not reference them anywhere else. We had them stored in a git-ignored filed in a local copy of this repo. Now we no longer need to keep this special copy of the repo locally. development and deployment can happen anywhere. just make sure the EVs are set in ELB