remove 'land' page, just load page and fetch data
previously, we were loading the landing page, which was essentially a buffering page just to show the user while we ran the selenium script, which could take while. then, once it finished, we would load a brand new page at a new route (live). now, instead of that, immediately return the bare minimum html needed, then use js to fetch the data from the server and display it once it is ready. in the meantime we'll display those pulsating dots to indicate that the data is loading in the background.

this eliminates the need for separate routes. also, previously, if you refreshed the 'live' route, it just sat there reloading the whole selenium script. now, if the user refreshes, we're sent back to the beginning with the loading dots, etc. better user experience.